timing.tests: Move SegmentStoreAnalysisTest to stubs directory
