+++ analysis/org.eclipse.tracecompass.analysis.timing.core.tests/.classpath
@@ -11,5 +11,6 @@
 		</attributes>
 	</classpathentry>
 	<classpathentry kind="src" path="src"/>
+	<classpathentry kind="src" path="stubs"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
