+++ analysis/org.eclipse.tracecompass.analysis.timing.core.tests/META-INF/MANIFEST.MF
@@ -23,7 +23,9 @@ Require-Bundle: org.junit;bundle-version="4.0.0",
 Import-Package: com.google.common.collect,
  org.eclipse.tracecompass.testtraces.ctf
 Export-Package: org.eclipse.tracecompass.analysis.timing.core.tests,
+ org.eclipse.tracecompass.analysis.timing.core.tests.segmentstore,
  org.eclipse.tracecompass.analysis.timing.core.tests.segmentstore.statistics,
- org.eclipse.tracecompass.analysis.timing.core.tests.statistics
+ org.eclipse.tracecompass.analysis.timing.core.tests.statistics,
+ org.eclipse.tracecompass.analysis.timing.core.tests.stubs.segmentstore
 Bundle-Activator: org.eclipse.tracecompass.analysis.timing.core.tests.Activator
 Automatic-Module-Name: org.eclipse.tracecompass.analysis.timing.core.tests
