+++ analysis/org.eclipse.tracecompass.analysis.timing.core.tests/build.properties
@@ -9,7 +9,8 @@
 # SPDX-License-Identifier: EPL-2.0
 ###############################################################################
 
-source.. = src/
+source.. = src/,\
+           stubs/
 output.. = bin/
 bin.includes = META-INF/,\
                .,\
